Conflicts:
	examples/vp9_spatial_svc_encoder.c
	examples/vpx_temporal_svc_encoder.c
	test/convolve_test.cc
	test/dct16x16_test.cc
	test/dct32x32_test.cc
	test/fdct4x4_test.cc
	test/fdct8x8_test.cc
	test/lpf_8_test.cc
	test/partial_idct_test.cc
	test/test.mk
	test/vp9_intrapred_test.cc
	tools_common.c
	vp8/vp8_cx_iface.c
	vp9/common/vp9_alloccommon.c
	vp9/common/vp9_common.h
	vp9/common/vp9_convolve.c
	vp9/common/vp9_convolve.h
	vp9/common/vp9_entropy.c
	vp9/common/vp9_entropy.h
	vp9/common/vp9_idct.c
	vp9/common/vp9_idct.h
	vp9/common/vp9_loopfilter.c
	vp9/common/vp9_loopfilter_filters.c
	vp9/common/vp9_onyxc_int.h
	vp9/common/vp9_postproc.c
	vp9/common/vp9_quant_common.c
	vp9/common/vp9_quant_common.h
	vp9/common/vp9_reconinter.c
	vp9/common/vp9_reconinter.h
	vp9/common/vp9_reconintra.c
	vp9/common/vp9_rtcd_defs.pl
	vp9/common/vp9_scale.c
	vp9/common/vp9_scale.h
	vp9/common/x86/vp9_asm_stubs.c
	vp9/common/x86/vp9_high_intrapred_sse2.asm
	vp9/common/x86/vp9_high_loopfilter_intrin_sse2.c
	vp9/common/x86/vp9_high_subpixel_8t_sse2.asm
	vp9/common/x86/vp9_high_subpixel_bilinear_sse2.asm
	vp9/decoder/vp9_decodeframe.c
	vp9/decoder/vp9_detokenize.c
	vp9/encoder/vp9_aq_complexity.c
	vp9/encoder/vp9_aq_cyclicrefresh.c
	vp9/encoder/vp9_aq_variance.c
	vp9/encoder/vp9_bitstream.c
	vp9/encoder/vp9_block.h
	vp9/encoder/vp9_context_tree.h
	vp9/encoder/vp9_dct.c
	vp9/encoder/vp9_encodeframe.c
	vp9/encoder/vp9_encodemb.c
	vp9/encoder/vp9_encoder.c
	vp9/encoder/vp9_encoder.h
	vp9/encoder/vp9_extend.c
	vp9/encoder/vp9_firstpass.c
	vp9/encoder/vp9_lookahead.c
	vp9/encoder/vp9_lookahead.h
	vp9/encoder/vp9_mcomp.c
	vp9/encoder/vp9_picklpf.c
	vp9/encoder/vp9_pickmode.c
	vp9/encoder/vp9_quantize.c
	vp9/encoder/vp9_quantize.h
	vp9/encoder/vp9_ratectrl.c
	vp9/encoder/vp9_rd.c
	vp9/encoder/vp9_rdopt.c
	vp9/encoder/vp9_resize.c
	vp9/encoder/vp9_resize.h
	vp9/encoder/vp9_sad.c
	vp9/encoder/vp9_ssim.c
	vp9/encoder/vp9_ssim.h
	vp9/encoder/vp9_temporal_filter.c
	vp9/encoder/vp9_tokenize.c
	vp9/encoder/vp9_tokenize.h
	vp9/encoder/vp9_variance.c
	vp9/encoder/vp9_variance.h
	vp9/vp9_common.mk
	vp9/vp9_cx_iface.c
	vp9/vp9_dx_iface.c
	vp9/vp9_iface_common.h
	vpx/src/svc_encodeframe.c
	vpx/src/vpx_image.c
	vpx/vp8dx.h
	vpx/vpx_codec.h
	vpx/vpx_encoder.h
	vpx/vpx_image.h
	vpx_mem/vpx_mem.c
	vpx_mem/vpx_mem.h
	vpx_scale/generic/yv12config.c
	vpx_scale/generic/yv12extend.c
	vpx_scale/yv12config.h
	vpxdec.c
	vpxenc.c

+++ vp8/common/arm/neon/bilinearpredict_neon.c
@@ -10,7 +10,7 @@
 
 #include <arm_neon.h>
 
-static const uint16_t bifilter4_coeff[8][2] = {
+static const uint8_t bifilter4_coeff[8][2] = {
     {128,   0},
     {112,  16},
     { 96,  32},
@@ -64,8 +64,8 @@ void vp8_bilinear_predict4x4_neon(
         q1u8 = vcombine_u8(d2u8, d3u8);
         q2u8 = vcombine_u8(d4u8, d5u8);
 
-        d0u8 = vdup_n_u8((uint8_t)bifilter4_coeff[xoffset][0]);
-        d1u8 = vdup_n_u8((uint8_t)bifilter4_coeff[xoffset][1]);
+        d0u8 = vdup_n_u8(bifilter4_coeff[xoffset][0]);
+        d1u8 = vdup_n_u8(bifilter4_coeff[xoffset][1]);
 
         q4u64  = vshrq_n_u64(vreinterpretq_u64_u8(q1u8), 8);
         q5u64  = vshrq_n_u64(vreinterpretq_u64_u8(q2u8), 8);
@@ -155,8 +155,8 @@ void vp8_bilinear_predict8x4_neon(
         q4u8 = vld1q_u8(src_ptr); src_ptr += src_pixels_per_line;
         q5u8 = vld1q_u8(src_ptr);
 
-        d0u8 = vdup_n_u8((uint8_t)bifilter4_coeff[xoffset][0]);
-        d1u8 = vdup_n_u8((uint8_t)bifilter4_coeff[xoffset][1]);
+        d0u8 = vdup_n_u8(bifilter4_coeff[xoffset][0]);
+        d1u8 = vdup_n_u8(bifilter4_coeff[xoffset][1]);
 
         q6u16 = vmull_u8(vget_low_u8(q1u8), d0u8);
         q7u16 = vmull_u8(vget_low_u8(q2u8), d0u8);
@@ -245,8 +245,8 @@ void vp8_bilinear_predict8x8_neon(
         q3u8 = vld1q_u8(src_ptr); src_ptr += src_pixels_per_line;
         q4u8 = vld1q_u8(src_ptr); src_ptr += src_pixels_per_line;
 
-        d0u8 = vdup_n_u8((uint8_t)bifilter4_coeff[xoffset][0]);
-        d1u8 = vdup_n_u8((uint8_t)bifilter4_coeff[xoffset][1]);
+        d0u8 = vdup_n_u8(bifilter4_coeff[xoffset][0]);
+        d1u8 = vdup_n_u8(bifilter4_coeff[xoffset][1]);
 
         q6u16 = vmull_u8(vget_low_u8(q1u8), d0u8);
         q7u16 = vmull_u8(vget_low_u8(q2u8), d0u8);
